Move concept views inside mapper directory
