+++ frontend/src/views/mapper/Course.js
@@ -4,7 +4,7 @@ import { makeStyles } from '@material-ui/core/styles'
 import { Button, Card, CardHeader, CardContent, List, IconButton } from '@material-ui/core'
 import { Edit as EditIcon } from '@material-ui/icons'
 
-import Concept from '../concept/Concept'
+import Concept from './concept/Concept'
 import { useLoginStateValue } from '../../store'
 
 const useStyles = makeStyles(theme => ({
